+++ examples/cvu/prg/prg.c
@@ -4,7 +4,7 @@
  * Purpose: Implementation file for the prg example project.
  *
  * Created: 8th December 2011
- * Updated: 2nd February 2024
+ * Updated: 12th July 2024
  *
  * ////////////////////////////////////////////////////////////////////// */
 
@@ -118,7 +118,7 @@ int main(int argc, char** argv)
     else
     {
       puts("");
-      printf("flags:\t%lu\n", args->numFlags);
+      printf("flags:\t%lu\n", (unsigned long)args->numFlags);
       { size_t i; for (i = 0; i != args->numFlags; ++i)
       {
           clasp_argument_t const* const flag = args->flags + i;
@@ -129,7 +129,7 @@ int main(int argc, char** argv)
 
 
       puts("");
-      printf("options:\t%lu\n", args->numOptions);
+      printf("options:\t%lu\n", (unsigned long)args->numOptions);
 
       { size_t i; for (i = 0; i != args->numOptions; ++i)
       {
@@ -141,7 +141,7 @@ int main(int argc, char** argv)
 
 
       puts("");
-      printf("values:\t%lu\n", args->numValues);
+      printf("values:\t%lu\n", (unsigned long)args->numValues);
 
       { size_t i; for (i = 0; i != args->numValues; ++i)
       {
